Move reusable tableDefinition into own component
